+++ ifinitySDK.framework/Versions/A/Headers/IFMBeacon+helper.h
@@ -12,7 +12,7 @@
 
 + (IFMBeacon *)fetchBeaconWithRemoteID:(NSNumber *)remoteID managedObjectContext:(NSManagedObjectContext *)context;
 + (NSArray *)fetchBeaconsWithFloorplanID:(NSNumber *)floorplanID managedObjectContext:(NSManagedObjectContext *)context;
-
++ (NSArray *)fetchBeaconsWithVenueId:(NSNumber *)venueId;
 + (IFMBeacon *)addBeaconRemoteID:(NSNumber *)remoteID
                        areaID:(NSNumber *)areaID
                   floorplanID:(NSNumber *)floorplanID