+++ ifinitySDK.framework/Versions/A/Headers/IFMContent.h
@@ -2,7 +2,7 @@
 //  IFMContent.h
 //  ifinitySDK
 //
-//  Created by Wojciech Chojnacki on 14.04.2015.
+//  Created by Mariusz Graczkowski on 14.05.2015.
 //  Copyright (c) 2015 GetIfinity. All rights reserved.
 //
 
@@ -17,9 +17,10 @@
 @property (nonatomic, retain) NSString * name;
 @property (nonatomic, retain) NSNumber * notification;
 @property (nonatomic, retain) NSNumber * remote_id;
+@property (nonatomic, retain) NSString * url;
 @property (nonatomic, retain) NSSet *areas;
-@property (nonatomic, retain) IFMVenue *venue;
 @property (nonatomic, retain) IFMContentImage *icon;
+@property (nonatomic, retain) IFMVenue *venue;
 @end
 
 @interface IFMContent (CoreDataGeneratedAccessors)